services/auspex/requirements.txt
ib_insync>=0.9.86
psycopg[binary]>=3.1
psycopg_pool>=3.1
nats-py>=2.6
redis>=5.0
structlog>=23.0
httpx>=0.25
